requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
black==20.8b1
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
dnd-character==23.7.29
exceptiongroup==1.2.0
fantasynames==0.1.2
flake8==3.9.2
Flask==3.0.2
Flask-Cors==4.0.0
flask_sqlalchemy==3.1.1
frozenlist==1.4.1
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.0
google-cloud-aiplatform==1.42.1
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-generativeai==0.3.2
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.60.1
grpcio-status==1.60.1
gunicorn==21.2.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.7
langchain-community==0.0.20
langchain-core==0.1.23
langchain-google-genai==0.0.9
langchain-google-vertexai==0.0.5
langsmith==0.0.87
MarkupSafe==2.1.5
marshmallow==3.20.2
mccabe==0.6.1
multidict==6.0.5
mypy==0.812
mypy-extensions==0.4.4
numpy==1.26.4
packaging==23.2
pathspec==0.12.1
pluggy==1.4.0
proto-plus==1.23.0
protobuf==4.25.3
py==1.11.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.7.0
pydantic==2.6.1
pydantic_core==2.16.2
pyflakes==2.3.1
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
shapely==2.0.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
tenacity==8.2.3
toml==0.10.2
tqdm==4.66.2
typed-ast==1.4.3
types-protobuf==4.24.0.20240129
types-requests==2.31.0.20240218
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.1
Werkzeug==3.0.1
yarl==1.9.4
